Introduce package extension for SciMLSensitivity
