+++ ext/SciMLSensitivityExtension.jl
@@ -1,0 +1,14 @@
+module SciMLSensitivityExtension
+
+using PEtab 
+using SciMLBase
+using ModelingToolkit
+using OrdinaryDiffEq
+using ForwardDiff
+using ReverseDiff
+using SciMLSensitivity
+
+include(joinpath(@__DIR__, "SciMLSensitivityExtension", "Adjoint.jl"))
+include(joinpath(@__DIR__, "SciMLSensitivityExtension", "Forward_equations.jl"))
+
+end
