+++ src/Derivatives/Forward_sensitivity_equations.jl
@@ -15,7 +15,7 @@ function computeGradientForwardEqDynamicθ!(gradient::Vector{Float64},
                                            θ_observable::Vector{Float64},
                                            θ_nonDynamic::Vector{Float64},
                                            petabModel::PEtabModel,
-                                           sensealg::Union{Symbol, SciMLSensitivity.AbstractForwardSensitivityAlgorithm},
+                                           sensealg,
                                            odeProblem::ODEProblem,
                                            simulationInfo::SimulationInfo,
                                            θ_indices::ParameterIndices,
@@ -65,25 +65,6 @@ function computeGradientForwardEqDynamicθ!(gradient::Vector{Float64},
 end
 
 
-function solveForSensitivites(odeProblem::ODEProblem,
-                              simulationInfo::SimulationInfo,
-                              θ_indices::ParameterIndices,
-                              petabModel::PEtabModel,
-                              sensealg::SciMLSensitivity.AbstractForwardSensitivityAlgorithm,
-                              θ_dynamic::AbstractVector,
-                              solveOdeModelAllConditions!::Function,
-                              cfg::Nothing,
-                              petabODECache::PEtabODEProblemCache,
-                              expIDSolve::Vector{Symbol},
-                              splitOverConditions::Bool,
-                              isRemade::Bool=false)
-
-    nModelStates = length(petabModel.stateNames)
-    _odeProblem = remake(odeProblem, p = convert.(eltype(θ_dynamic), odeProblem.p), u0 = convert.(eltype(θ_dynamic), odeProblem.u0))
-    changeODEProblemParameters!(_odeProblem.p, (@view _odeProblem.u0[1:nModelStates]), θ_dynamic, θ_indices, petabModel)
-    success = solveOdeModelAllConditions!(simulationInfo.odeSolutionsDerivatives, _odeProblem, θ_dynamic, expIDSolve)
-    return success
-end
 function solveForSensitivites(odeProblem::ODEProblem,
                               simulationInfo::SimulationInfo,
                               θ_indices::ParameterIndices,
@@ -163,57 +144,6 @@ function solveForSensitivites(odeProblem::ODEProblem,
 end
 
 
-function computeGradientForwardExpCond!(gradient::Vector{Float64},
-                                        sol::ODESolution,
-                                        petabODECache::PEtabODEProblemCache,
-                                        sensealg::SciMLSensitivity.AbstractForwardSensitivityAlgorithm,
-                                        θ_dynamic::Vector{Float64},
-                                        θ_sd::Vector{Float64},
-                                        θ_observable::Vector{Float64},
-                                        θ_nonDynamic::Vector{Float64},
-                                        experimentalConditionId::Symbol,
-                                        simulationConditionId::Symbol,
-                                        simulationInfo::SimulationInfo,
-                                        petabModel::PEtabModel,
-                                        θ_indices::ParameterIndices,
-                                        measurementInfo::MeasurementsInfo,
-                                        parameterInfo::ParametersInfo)
-
-    iPerTimePoint = simulationInfo.iPerTimePoint[experimentalConditionId]
-    timeObserved = simulationInfo.timeObserved[experimentalConditionId]
-
-    # To compute
-    compute∂G∂u = (out, u, p, t, i) -> begin compute∂G∂_(out, u, p, t, i, iPerTimePoint,
-                                                         measurementInfo, parameterInfo,
-                                                         θ_indices, petabModel,
-                                                         θ_sd, θ_observable, θ_nonDynamic,
-                                                         petabODECache.∂h∂u, petabODECache.∂σ∂u, compute∂G∂U=true)
-                                            end
-    compute∂G∂p = (out, u, p, t, i) -> begin compute∂G∂_(out, u, p, t, i, iPerTimePoint,
-                                                         measurementInfo, parameterInfo,
-                                                         θ_indices, petabModel,
-                                                         θ_sd, θ_observable, θ_nonDynamic,
-                                                         petabODECache.∂h∂p, petabODECache.∂σ∂p, compute∂G∂U=false)
-                                        end
-
-    # Loop through solution and extract sensitivites
-    p = sol.prob.p
-    ∂G∂p, ∂G∂p_ = zeros(Float64, length(p)), zeros(Float64, length(p))
-    ∂G∂u = zeros(Float64, length(petabModel.stateNames))
-    _gradient = zeros(Float64, length(p))
-    for i in eachindex(timeObserved)
-        u, _S = extract_local_sensitivities(sol, i, true)
-        compute∂G∂u(∂G∂u, u, p, timeObserved[i], i)
-        compute∂G∂p(∂G∂p_, u, p, timeObserved[i], i)
-        _gradient .+= transpose(_S)*∂G∂u
-        ∂G∂p .+= ∂G∂p_
-    end
-
-    # Thus far have have computed dY/dθ, but for parameters on the log-scale we want dY/dθ_log. We can adjust via;
-    # dY/dθ_log = log(10) * θ * dY/dθ
-    adjustGradientTransformedParameters!(gradient, _gradient, ∂G∂p, θ_dynamic, θ_indices,
-                                         simulationConditionId, autoDiffSensitivites=false)
-end
 function computeGradientForwardExpCond!(gradient::Vector{Float64},
                                         sol::ODESolution,
                                         petabODECache::PEtabODEProblemCache,
