+++ src/Derivatives/Gradient.jl
@@ -148,7 +148,7 @@ function computeGradientForwardEquations!(gradient::Vector{Float64},
                                           computeCostNotODESystemθ::Function,
                                           petabModel::PEtabModel,
                                           odeProblem::ODEProblem,
-                                          sensealg::Union{Symbol, SciMLSensitivity.AbstractForwardSensitivityAlgorithm},
+                                          sensealg,
                                           simulationInfo::SimulationInfo,
                                           θ_indices::ParameterIndices,
                                           measurementInfo::MeasurementsInfo,
@@ -177,54 +177,6 @@ function computeGradientForwardEquations!(gradient::Vector{Float64},
                                       sensealg, odeProblem, simulationInfo, θ_indices, measurementInfo, parameterInfo,
                                       solveOdeModelAllConditions!, cfg, petabODECache, expIDSolve=expIDSolve,
                                       splitOverConditions=splitOverConditions, isRemade=isRemade)
-    @views gradient[θ_indices.iθ_dynamic] .= petabODECache.gradientDyanmicθ
-
-    # Happens when at least one forward pass fails and I set the gradient to 1e8
-    if !isempty(petabODECache.gradientDyanmicθ) && all(petabODECache.gradientDyanmicθ .== 0.0)
-        gradient .= 0.0
-        return
-    end
-
-    θ_notOdeSystem = @view θ_est[θ_indices.iθ_notOdeSystem]
-    ReverseDiff.gradient!(petabODECache.gradientNotODESystemθ, computeCostNotODESystemθ, θ_notOdeSystem)
-    @views gradient[θ_indices.iθ_notOdeSystem] .= petabODECache.gradientNotODESystemθ
-
-    if priorInfo.hasPriors == true
-        computeGradientPrior!(gradient, θ_est, θ_indices, priorInfo)
-    end
-end
-
-
-# Compute gradient via adjoint sensitivity analysis
-function computeGradientAdjointEquations!(gradient::Vector{Float64},
-                                          θ_est::Vector{Float64},
-                                          solverOptions::ODESolverOptions,
-                                          ssSolverOptions::SteadyStateSolverOptions,
-                                          computeCostNotODESystemθ::Function,
-                                          sensealg::SciMLSensitivity.AbstractAdjointSensitivityAlgorithm,
-                                          sensealgSS::SciMLSensitivity.AbstractAdjointSensitivityAlgorithm,
-                                          odeProblem::ODEProblem,
-                                          petabModel::PEtabModel,
-                                          simulationInfo::SimulationInfo,
-                                          θ_indices::ParameterIndices,
-                                          measurementInfo::MeasurementsInfo,
-                                          parameterInfo::ParametersInfo,
-                                          priorInfo::PriorInfo,
-                                          petabODECache::PEtabODEProblemCache,
-                                          petabODESolverCache::PEtabODESolverCache;
-                                          expIDSolve::Vector{Symbol} = [:all])
-
-    splitParameterVector!(θ_est, θ_indices, petabODECache)
-    θ_dynamic = petabODECache.θ_dynamic
-    θ_observable = petabODECache.θ_observable
-    θ_sd = petabODECache.θ_sd
-    θ_nonDynamic = petabODECache.θ_nonDynamic
-
-    # Calculate gradient seperately for dynamic and non dynamic parameter.
-    computeGradientAdjointDynamicθ(petabODECache.gradientDyanmicθ, θ_dynamic, θ_sd, θ_observable, θ_nonDynamic, odeProblem, solverOptions,
-                                   ssSolverOptions, sensealg, petabModel, simulationInfo, θ_indices, measurementInfo, parameterInfo,
-                                   petabODECache, petabODESolverCache; expIDSolve=expIDSolve,
-                                   sensealgSS=sensealgSS)
     @views gradient[θ_indices.iθ_dynamic] .= petabODECache.gradientDyanmicθ
 
     # Happens when at least one forward pass fails and I set the gradient to 1e8
