+++ src/PEtab_select/PEtab_select.jl
@@ -28,8 +28,8 @@ function runPEtabSelect(pathYAML::String,
                         gradientMethod::Union{Nothing, Symbol}=nothing,
                         hessianMethod::Union{Nothing, Symbol}=nothing,
                         sparseJacobian::Union{Nothing, Bool}=nothing,
-                        sensealg::Union{Nothing, Symbol, SciMLBase.AbstractSensitivityAlgorithm}=nothing,
-                        sensealgSS::Union{Nothing, SciMLSensitivity.AbstractAdjointSensitivityAlgorithm}=nothing,
+                        sensealg=nothing,
+                        sensealgSS=nothing,
                         chunkSize::Union{Nothing, Int64}=nothing,
                         splitOverConditions::Bool=false,
                         reuseS::Bool=false) where T <: QuasiMonteCarlo.SamplingAlgorithm
