+++ src/Process_PEtab_files/Get_simulation_info.jl
@@ -7,7 +7,7 @@
 
 function processSimulationInfo(petabModel::PEtabModel,
                                measurementInfo::MeasurementsInfo;
-                               sensealg::Union{Nothing, Symbol, SciMLSensitivity.AbstractForwardSensitivityAlgorithm, SciMLSensitivity.AbstractAdjointSensitivityAlgorithm}=InterpolatingAdjoint())::SimulationInfo
+                               sensealg)::SimulationInfo
 
     # An experimental Id is uniqely defined by a Pre-equlibrium- and Simulation-Id, where the former can be
     # empty. For each experimental ID we store three indices, i) preEqulibriumId, ii) simulationId and iii)
@@ -85,10 +85,6 @@ function processSimulationInfo(petabModel::PEtabModel,
     trackedCallbacks = Dict{Symbol, SciMLBase.DECallback}()
     for name in experimentalConditionId
         callbacks[name] = deepcopy(petabModel.modelCallbackSet)
-    end
-
-    if typeof(sensealg) <: Symbol || isnothing(sensealg)
-        sensealg = InterpolatingAdjoint()
     end
 
     simulationInfo = SimulationInfo(preEquilibrationConditionId,
