+++ src/Solve_ODE/Helper.jl
@@ -36,17 +36,10 @@ end
 function getCallbackSet(odeProblem::ODEProblem,
                         simulationInfo::SimulationInfo,
                         simulationConditionId::Symbol,
-                        trackCallback::Bool)::SciMLBase.DECallback
+                        sensealg)::SciMLBase.DECallback
 
-    if trackCallback == true
-        cbSet = SciMLSensitivity.track_callbacks(simulationInfo.callbacks[simulationConditionId], odeProblem.tspan[1],
-                                                 odeProblem.u0, odeProblem.p, simulationInfo.sensealg)
-        simulationInfo.trackedCallbacks[simulationConditionId] = cbSet
-        return cbSet
-    end
-    return simulationInfo.callbacks[simulationConditionId]
+    return simulationInfo.callbacks[simulationConditionId]                        
 end
-
 
 
 function setTspanODEProblem(odeProblem::ODEProblem,
